+++ scout/cors.py
