+++ scout/static/index.html
