+++ scout/templates/index.html
