+++ scout/templates/login.html
