--- server/static/js/templates.js
@@ -1,2 +1,0 @@
-Ember.TEMPLATES.application=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){e.buffer.push('\n        <div class="menu__item__icon"><img src="/static/img/icon_8199.svg"></div>\n        <div class="menu__item__label">Families</div>\n      ')}function l(s,e){e.buffer.push('\n        <div class="menu__item__icon"><img src="/static/img/icon_19741.svg"></div>\n        <div class="menu__item__label">Settings</div>\n      ')}function i(s,e){e.buffer.push('\n        <div class="menu__item__icon"><img src="/static/img/icon_4240.svg"></div>\n        <div class="menu__item__label">Issue</div>\n      ')}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var r,p,c,u="",o=this,d=a.helperMissing;return n.buffer.push('\n<div class="panel__wrapper">\n\n  <div class="panel--menu">\n    <div class="panel__menu--footer">'),r=a._triageMustache.call(e,"user.email",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(r||0===r)&&n.buffer.push(r),n.buffer.push('</div>\n\n    <div class="panel__body">\n      '),p=a["link-to"]||e&&e["link-to"],c={hash:{classNames:"menu__item"},hashTypes:{classNames:"STRING"},hashContexts:{classNames:e},inverse:o.noop,fn:o.program(1,h,n),contexts:[e],types:["STRING"],data:n},r=p?p.call(e,"index",c):d.call(e,"link-to","index",c),(r||0===r)&&n.buffer.push(r),n.buffer.push("\n      "),p=a["link-to"]||e&&e["link-to"],c={hash:{classNames:"menu__item"},hashTypes:{classNames:"STRING"},hashContexts:{classNames:e},inverse:o.noop,fn:o.program(3,l,n),contexts:[e],types:["STRING"],data:n},r=p?p.call(e,"settings",c):d.call(e,"link-to","settings",c),(r||0===r)&&n.buffer.push(r),n.buffer.push("\n      "),p=a["link-to"]||e&&e["link-to"],c={hash:{classNames:"menu__item"},hashTypes:{classNames:"STRING"},hashContexts:{classNames:e},inverse:o.noop,fn:o.program(5,i,n),contexts:[e],types:["STRING"],data:n},r=p?p.call(e,"issue",c):d.call(e,"link-to","issue",c),(r||0===r)&&n.buffer.push(r),n.buffer.push('\n      <a href="/logout" class="menu__item">\n        <div class="menu__item__icon"><img src="/static/img/icon_4930.svg"></div>\n        <div class="menu__item__label">Logout</div>\n      </a>\n    </div>\n  </div>\n\n  '),r=a._triageMustache.call(e,"outlet",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(r||0===r)&&n.buffer.push(r),n.buffer.push("\n\n</div>\n"),u}),Ember.TEMPLATES.dashboard=Ember.Handlebars.template(function(s,e,a,t,n){this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{},n.buffer.push('<div class="panel variant">\n\n  <div class="panel__overlay--header">\n    <div class="panel__overlay__title">Dashboard</div>\n  </div>\n\n  <div class="panel__body">\n\n    <div class="box">\n      <h3>Family 1</h3>\n\n      <div class="progress__wrapper">\n        <div class="progress__bar"></div>\n\n        <div class="progress__well">\n          <div class="progress__mark">\n            <div class="progress__bar__dot"></div>\n            <div class="progress__mark__label">Sequenced</div>\n          </div>\n          <div class="progress__mark">\n            <div class="progress__bar__dot"></div>\n            <div class="progress__mark__label">Candidate gene</div>\n          </div>\n          <div class="progress__mark">\n            <div class="progress__bar__dot"></div>\n            <div class="progress__mark__label">Sanger ordered</div>\n          </div>\n          <div class="progress__mark">\n            <div class="progress__bar__dot"></div>\n            <div class="progress__mark__label">Solved</div>\n          </div>\n        </div>\n      </div>        \n    </div>\n\n  </div>\n</div>\n')}),Ember.TEMPLATES.families=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){e.buffer.push('\n    <div class="panel__overlay__column">NGS update</div>\n    ')}function l(s,e){var t,n,h,l="";return e.buffer.push("\n        "),n=a["link-to"]||s&&s["link-to"],h={hash:{classNames:"list__item"},hashTypes:{classNames:"STRING"},hashContexts:{classNames:s},inverse:f.noop,fn:f.program(4,i,e),contexts:[s,s],types:["STRING","ID"],data:e},t=n?n.call(s,"family","family",h):o.call(s,"link-to","family","family",h),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n      "),l}function i(s,e){var t,n="";return e.buffer.push('\n\n          <div class="list__item__content">\n            <div class="list__item__content__text">\n              '),t=a.unless.call(s,"familyLoaded",{hash:{},hashTypes:{},hashContexts:{},inverse:f.noop,fn:f.program(5,r,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n              "),t=a._triageMustache.call(s,"family.id",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n            </div>\n\n            "),t=a.unless.call(s,"familyLoaded",{hash:{},hashTypes:{},hashContexts:{},inverse:f.noop,fn:f.program(7,p,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n          </div>\n\n        "),n}function r(s,e){e.buffer.push("\n                Family\n              ")}function p(s,e){var t,n,h="";return e.buffer.push('\n            <div class="list__item__content__label">\n              '),e.buffer.push(d((t=a.fromNow||s&&s.fromNow,n={hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e},t?t.call(s,"family.analyzedDate",n):o.call(s,"fromNow","family.analyzedDate",n)))),e.buffer.push("\n            </div>\n            "),h}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var c,u="",o=a.helperMissing,d=this.escapeExpression,f=this;return n.buffer.push("<div "),n.buffer.push(d(a["bind-attr"].call(e,{hash:{"class":"familyLoaded:panel--sidebar:panel :families"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n\n  <div class="panel__overlay--header">\n    <div class="panel__overlay__title">Families</div>\n    \n    <div class="panel__overlay__column">ID</div>\n    '),c=a.unless.call(e,"familyLoaded",{hash:{},hashTypes:{},hashContexts:{},inverse:f.noop,fn:f.program(1,h,n),contexts:[e],types:["ID"],data:n}),(c||0===c)&&n.buffer.push(c),n.buffer.push('\n  </div>\n\n  <div class="panel__body">\n    <div class="list--vertical">\n      '),c=a.each.call(e,"family","in","model",{hash:{},hashTypes:{},hashContexts:{},inverse:f.noop,fn:f.program(3,l,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(c||0===c)&&n.buffer.push(c),n.buffer.push("\n    </div>\n  </div>\n\n</div>\n\n"),c=a._triageMustache.call(e,"outlet",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(c||0===c)&&n.buffer.push(c),n.buffer.push("\n"),u}),Ember.TEMPLATES.family=Ember.Handlebars.template(function(s,e,a,t,n){this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var h,l="";return h=a._triageMustache.call(e,"outlet",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(h||0===h)&&n.buffer.push(h),n.buffer.push("\n"),l}),Ember.TEMPLATES["gt-call"]=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){var t,n="";return e.buffer.push('\n    <th class="predict__part">'),t=a._triageMustache.call(s,"idn",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</th>\n    "),n}function l(s,e){var t,n,h,l="";return e.buffer.push('\n    <tr>\n    <td class="predict__part center-text">\n      '),n=a["link-to"]||s&&s["link-to"],h={hash:{},hashTypes:{},hashContexts:{},inverse:u.noop,fn:u.program(4,i,e),contexts:[s,s],types:["STRING","ID"],data:e},t=n?n.call(s,"variant","compound.vpk",h):o.call(s,"link-to","variant","compound.vpk",h),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n    </td>\n\n    "),t=a.each.call(s,"call","in","compound.gtCalls",{hash:{},hashTypes:{},hashContexts:{},inverse:u.noop,fn:u.program(6,r,e),contexts:[s,s,s],types:["ID","ID","ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n\n    <td class="predict__part">'),t=a._triageMustache.call(s,"compound.gene_annotation",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</td>\n    <td class="predict__part">'),t=a._triageMustache.call(s,"compound.functional_annotation",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</td>\n    <td class="predict__part">'),t=a._triageMustache.call(s,"compound.gene_model",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</td>\n    </tr>\n  "),l}function i(s,e){var t,n="";return e.buffer.push("\n        "),t=a._triageMustache.call(s,"compound.rank_score",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n      "),n}function r(s,e){var t,n="";return e.buffer.push('\n    <td class="predict__part">'),t=a._triageMustache.call(s,"call",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</td>\n    "),n}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var p,c="",u=this,o=a.helperMissing;return n.buffer.push('<table style="width: 100%">\n  <thead>\n  <tr>\n    <th class="predict__part">Rank score</th>\n    '),p=a.each.call(e,"idn","in","gt.compounds.1.idns",{hash:{},hashTypes:{},hashContexts:{},inverse:u.noop,fn:u.program(1,h,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(p||0===p)&&n.buffer.push(p),n.buffer.push('\n    <th class="predict__part">Gene annotation</th>\n    <th class="predict__part">Func. annotation</th>\n    <th class="predict__part">Inheritance</th>\n  </tr>\n  </thead>\n\n  '),p=a.each.call(e,"compound","in","gt.compounds",{hash:{},hashTypes:{},hashContexts:{},inverse:u.noop,fn:u.program(3,l,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(p||0===p)&&n.buffer.push(p),n.buffer.push("\n</table>"),c}),Ember.TEMPLATES.index=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){var t,n="";return e.buffer.push("\n      "),t=a.unless.call(s,"family.isHidden",{hash:{},hashTypes:{},hashContexts:{},inverse:u.noop,fn:u.program(2,l,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n      "),n}function l(s,e){var t,n,h,l="";return e.buffer.push('\n        <div class="list__item">\n          <div class="list__column">\n            <div class="btn-group">\n              <div class="btn">\n                '),n=a["link-to"]||s&&s["link-to"],h={hash:{},hashTypes:{},hashContexts:{},inverse:u.noop,fn:u.program(3,i,e),contexts:[s,s],types:["STRING","ID"],data:e},t=n?n.call(s,"family","family",h):o.call(s,"link-to","family","family",h),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n              </div>\n              <div class="btn">\n                '),n=a["link-to"]||s&&s["link-to"],h={hash:{},hashTypes:{},hashContexts:{},inverse:u.noop,fn:u.program(5,r,e),contexts:[s,s],types:["STRING","ID"],data:e},t=n?n.call(s,"variants","family",h):o.call(s,"link-to","variants","family",h),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n              </div>\n              <div "),e.buffer.push(d(a.action.call(s,"hideFamily","family",{hash:{},hashTypes:{},hashContexts:{},contexts:[s,s],types:["STRING","ID"],data:e}))),e.buffer.push(' class="btn">\n                <div class="btn__body">Hide</div>\n              </div>\n            </div>\n          </div>\n          \n          <div class="list__column right-text">\n            '),e.buffer.push(d((n=a.fromNow||s&&s.fromNow,h={hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e},n?n.call(s,"family.analyzedDate",h):o.call(s,"fromNow","family.analyzedDate",h)))),e.buffer.push("\n          </div>\n        </div>\n      "),l}function i(s,e){var t,n="";return e.buffer.push('\n                  <div class="btn__body">Family '),t=a._triageMustache.call(s,"family.id",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n                "),n}function r(s,e){e.buffer.push('\n                  <div class="btn__body">Variants</div>\n                ')}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var p,c="",u=this,o=a.helperMissing,d=this.escapeExpression;return n.buffer.push('<div class="panel families">\n\n  <div class="panel__overlay--header">\n    <div class="panel__overlay__title">Families</div>\n    \n    <div class="list--vertical">\n      <div class="list__item">\n        <div class="list__column">Links</div>\n        <div class="list__column right-text">NGS update</div>\n      </div>\n    </div>\n  </div>\n\n  <div class="panel__body">\n    <div class="list--vertical">\n      '),p=a.each.call(e,"family","in","model",{hash:{},hashTypes:{},hashContexts:{},inverse:u.noop,fn:u.program(1,h,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(p||0===p)&&n.buffer.push(p),n.buffer.push("\n    </div>\n  </div>\n\n</div>\n\n"),p=a._triageMustache.call(e,"outlet",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(p||0===p)&&n.buffer.push(p),n.buffer.push("\n"),c}),Ember.TEMPLATES.issue=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){var t,n,h,l="";return e.buffer.push('\n      <div class="list__item">\n        <div class="list__column fixed">\n          <a '),e.buffer.push(b(a["bind-attr"].call(s,{hash:{href:"issue.url"},hashTypes:{href:"STRING"},hashContexts:{href:s},contexts:[],types:[],data:e}))),e.buffer.push(' target="_blank">'),t=a._triageMustache.call(s,"issue.title",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</a>\n        </div>\n        <div class="list__column--small fixed">\n          '),e.buffer.push(b((n=a.fromNow||s&&s.fromNow,h={hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e},n?n.call(s,"issue.createdAt",h):v.call(s,"fromNow","issue.createdAt",h)))),e.buffer.push("\n        </div>\n      </div>\n      "),l}function l(s,e){e.buffer.push('\n      <div class="center-center no-comments"><small>No open issues.</small></div>\n      ')}function i(s,e){var t,n="";return e.buffer.push("\n\n    "),t=a["if"].call(s,"newIssue.isSaving",{hash:{},hashTypes:{},hashContexts:{},inverse:m.program(8,p,e),fn:m.program(6,r,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n\n  "),n}function r(s,e){e.buffer.push('\n    <div class="issue-sending">Sending issue...</div>\n\n    ')}function p(s,e){var t,n,h,l="";return e.buffer.push('\n    <form class="issue-form" '),e.buffer.push(b(a.action.call(s,"confirmIssue",{hash:{on:"submit"},hashTypes:{on:"STRING"},hashContexts:{on:s},contexts:[s],types:["STRING"],data:e}))),e.buffer.push(">\n      <h4>Submit a new issue</h4>\n      "),e.buffer.push(b((n=a.input||s&&s.input,h={hash:{value:"newIssue.title",type:"text",placeholder:"Title",classNames:"text-input"},hashTypes:{value:"ID",type:"STRING",placeholder:"STRING",classNames:"STRING"},hashContexts:{value:s,type:s,placeholder:s,classNames:s},contexts:[],types:[],data:e},n?n.call(s,h):v.call(s,"input",h)))),e.buffer.push("\n      "),e.buffer.push(b((n=a.textarea||s&&s.textarea,h={hash:{value:"newIssue.body",placeholder:"Describe your problem/request",classNames:"text-input"},hashTypes:{value:"ID",placeholder:"STRING",classNames:"STRING"},hashContexts:{value:s,placeholder:s,classNames:s},contexts:[],types:[],data:e},n?n.call(s,h):v.call(s,"textarea",h)))),e.buffer.push("\n\n      "),t=a.unless.call(s,"isConfirmingIssue",{hash:{},hashTypes:{},hashContexts:{},inverse:m.program(11,u,e),fn:m.program(9,c,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n    </form>\n    "),l}function c(s,e){e.buffer.push('\n        <button type="submit" class="btn">\n          <div class="btn__body">Submit</div>\n        </button>\n      ')}function u(s,e){var t="";return e.buffer.push("\n        <div "),e.buffer.push(b(a.action.call(s,"createIssue",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push(' class="btn">\n          <div class="btn__body">Are you sure?</div>\n        </div>\n      '),t}function o(s,e){var t,n="";return e.buffer.push('\n\n    \n    <div class="issue-form">\n      <h4>Your issue</h4>\n      <p>Click the title to provide additional information and upload a screenshot.</p>\n      <a '),e.buffer.push(b(a["bind-attr"].call(s,{hash:{href:"message.html_url"},hashTypes:{href:"STRING"},hashContexts:{href:s},contexts:[],types:[],data:e}))),e.buffer.push(' class="issue-title">\n        '),t=a._triageMustache.call(s,"message.title",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n      </a>\n      <div class="issue-body-wrapper">'),e.buffer.push(b(a._triageMustache.call(s,"message.html",{hash:{unescaped:"true"},hashTypes:{unescaped:"STRING"},hashContexts:{unescaped:s},contexts:[s],types:["ID"],data:e}))),e.buffer.push("</div>\n\n      <div "),e.buffer.push(b(a.action.call(s,"reset",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push(' class="btn">\n        <div class="btn__body">New issue</div>\n      </div>\n    </div>    \n\n  '),n}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var d,f="",b=this.escapeExpression,v=a.helperMissing,m=this;return n.buffer.push('<div class="panel--sidebar issue-sidebar">\n  <div class="panel__body">\n    <div class="list--vertical">\n      '),d=a.each.call(e,"issue","in","model",{hash:{},hashTypes:{},hashContexts:{},inverse:m.program(3,l,n),fn:m.program(1,h,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(d||0===d)&&n.buffer.push(d),n.buffer.push('\n    </div>\n  </div>\n</div>\n\n<div class="panel">\n\n  <div class="center-center">\n  '),d=a["if"].call(e,"newIssue.isNew",{hash:{},hashTypes:{},hashContexts:{},inverse:m.program(13,o,n),fn:m.program(5,i,n),contexts:[e],types:["ID"],data:n}),(d||0===d)&&n.buffer.push(d),n.buffer.push("\n  </div>\n\n</div>\n"),f}),Ember.TEMPLATES["order-modal"]=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){var t,n,h="";return e.buffer.push('\n  <h5>Order Sanger sequencing</h5>\n\n  <form class="form--complex">\n    <label for="famil-id">Family</label>\n    <div class="form__input-wrapper">\n      '),e.buffer.push(c((t=a.input||s&&s.input,n={hash:{valueBinding:"family_id",id:"famil-id",classNames:"text-input"},hashTypes:{valueBinding:"STRING",id:"STRING",classNames:"STRING"},hashContexts:{valueBinding:s,id:s,classNames:s},contexts:[],types:[],data:e},t?t.call(s,n):p.call(s,"input",n)))),e.buffer.push('\n    </div>\n\n    <label for="hgnc-symbol">HGNC symbol</label>\n    <div class="form__input-wrapper">\n      '),e.buffer.push(c((t=a.input||s&&s.input,n={hash:{valueBinding:"hgnc_symbol",id:"hgnc-symbol",classNames:"text-input"},hashTypes:{valueBinding:"STRING",id:"STRING",classNames:"STRING"},hashContexts:{valueBinding:s,id:s,classNames:s},contexts:[],types:[],data:e},t?t.call(s,n):p.call(s,"input",n)))),e.buffer.push('\n    </div>\n\n    <label for="chr-pos">Chromosome position</label>\n    <div class="form__input-wrapper">\n      '),e.buffer.push(c((t=a.input||s&&s.input,n={hash:{valueBinding:"chr_pos",id:"chr-pos",classNames:"text-input"},hashTypes:{valueBinding:"STRING",id:"STRING",classNames:"STRING"},hashContexts:{valueBinding:s,id:s,classNames:s},contexts:[],types:[],data:e},t?t.call(s,n):p.call(s,"input",n)))),e.buffer.push('\n    </div>\n\n    <label for="amino-change">Amino acid change</label>\n    <div class="form__input-wrapper">\n      '),e.buffer.push(c((t=a.textarea||s&&s.textarea,n={hash:{valueBinding:"amino_change",id:"amino-change",classNames:"text-input"},hashTypes:{valueBinding:"STRING",id:"STRING",classNames:"STRING"},hashContexts:{valueBinding:s,id:s,classNames:s},contexts:[],types:[],data:e},t?t.call(s,n):p.call(s,"textarea",n)))),e.buffer.push('\n    </div>\n\n    <label for="gt-call">GT Call</label>\n    <div class="form__input-wrapper">\n      '),e.buffer.push(c((t=a.textarea||s&&s.textarea,n={hash:{valueBinding:"gt_call",id:"gt-call",classNames:"text-input"},hashTypes:{valueBinding:"STRING",id:"STRING",classNames:"STRING"},hashContexts:{valueBinding:s,id:s,classNames:s},contexts:[],types:[],data:e},t?t.call(s,n):p.call(s,"textarea",n)))),e.buffer.push('\n    </div>\n\n    <div class="btn-group">\n      <button '),e.buffer.push(c(a.action.call(s,"sendForm",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push(' class="btn">\n        <div class="btn__body">Confirm</div>\n      </button>\n      <button '),e.buffer.push(c(a.action.call(s,"close",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push(' class="btn">\n        <div class="btn__body">Cancel</div>\n      </button>\n    </div>\n  </form>\n'),h}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var l,i,r,p=a.helperMissing,c=this.escapeExpression,u=this;i=a["modal-dialog"]||e&&e["modal-dialog"],r={hash:{action:"close"},hashTypes:{action:"STRING"},hashContexts:{action:e},inverse:u.noop,fn:u.program(1,h,n),contexts:[],types:[],data:n},l=i?i.call(e,r):p.call(e,"modal-dialog",r),n.buffer.push(l||0===l?l:"")}),Ember.TEMPLATES.settings=Ember.Handlebars.template(function(s,e,a,t,n){this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var h="",l=this.escapeExpression;return n.buffer.push('<div class="panel settings">\n  <div class="panel__body settings-body">\n    <div class="center-center box">\n      <div class="btn-group">\n        <div '),n.buffer.push(l(a.action.call(e,"resetHidden","family",{hash:{},hashTypes:{},hashContexts:{},contexts:[e,e],types:["STRING","STRING"],data:n}))),n.buffer.push(' class="btn">\n          <div class="btn__body">Reset hidden families</div>\n        </div>\n        <div '),n.buffer.push(l(a.action.call(e,"resetHidden","variant",{hash:{},hashTypes:{},hashContexts:{},contexts:[e,e],types:["STRING","STRING"],data:n}))),n.buffer.push(' class="btn">\n          <div class="btn__body">Reset hidden variants</div>\n        </div>\n      </div>\n    </div>\n  </div>\n</div>\n'),h}),Ember.TEMPLATES.variant=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){var t,n,h="";return e.buffer.push("\n            "),e.buffer.push(D((t=a.capitalize||s&&s.capitalize,n={hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e},t?t.call(s,"hgncApprovedName",n):N.call(s,"capitalize","hgncApprovedName",n)))),e.buffer.push("\n          "),h}function l(s,e){var t,n="";return e.buffer.push("\n            "),t=a._triageMustache.call(s,"hgncSymbol",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push(" <small>"),t=a._triageMustache.call(s,"hgncSynonymsString",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</small>\n          "),n}function i(s,e){var t,n="";return e.buffer.push(" - "),t=a._triageMustache.call(s,"stopBp",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),n}function r(s,e){var t,n,h="";return e.buffer.push("\n              "),e.buffer.push(D((t=a.capitalize||s&&s.capitalize,n={hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e},t?t.call(s,"functionalAnnotation",n):N.call(s,"capitalize","functionalAnnotation",n)))),e.buffer.push("\n            "),h}function p(s,e){e.buffer.push("\n              No functional annotation\n            ")}function c(s,e){e.buffer.push("All variants")}function u(s,e){var t="";return e.buffer.push("\n            <div "),e.buffer.push(D(a.action.call(s,"hideInList",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push(' class="btn">\n              <div class="btn__body">Hide</div>\n            </div>\n            '),t}function o(s,e){var t="";return e.buffer.push("\n            <div "),e.buffer.push(D(a.action.call(s,"unhideInList",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push(' class="btn">\n              <div class="btn__body">Unhide</div>\n            </div>\n            '),t}function d(s,e){var t,n="";return e.buffer.push("\n                    <div>"),t=a._triageMustache.call(s,"syn",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n                  "),n}function f(s,e){var t,n="";return e.buffer.push('\n              <div class="list__item">\n                <div class="list__column--slim">'),t=a._triageMustache.call(s,"item",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n              </div>\n              "),n}function b(s,e){e.buffer.push('\n                <div class="center-center no-comments"><small>No functions annotated.</small></div>\n              ')}function v(s,e){var t,n="";return e.buffer.push('\n          <tr>\n            <td class="predict__main predict__part">'),t=a._triageMustache.call(s,"call.idn",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</td>\n            <td class="predict__part">'),t=a._triageMustache.call(s,"call.gt",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</td>\n            <td class="predict__part">'),t=a._triageMustache.call(s,"call.ad",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</td>\n            <td class="predict__part">'),t=a._triageMustache.call(s,"call.pl",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</td>\n            <td class="predict__part">'),t=a._triageMustache.call(s,"call.gq",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</td>\n          </tr>\n        "),n}function m(s,e){var t,n="";return e.buffer.push('\n            <td class="predict__part">'),t=a._triageMustache.call(s,"model",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</td>\n          "),n}function _(s,e){var t,n,h="";return e.buffer.push('\n    <div class="box compounds">\n      '),e.buffer.push(D((t=a.render||s&&s.render,n={hash:{},hashTypes:{},hashContexts:{},contexts:[s,s],types:["STRING","ID"],data:e},t?t.call(s,"gt-call","",n):N.call(s,"render","gt-call","",n)))),e.buffer.push("\n    </div>\n    "),h}function y(s,e){var t,n="";return e.buffer.push('\n            <div class="comment-box__wrapper">\n              <div class="comment-box">\n                <div class="comment-box__header">\n                  <div class="comment-box__header__body">\n                    <b>'),t=a._triageMustache.call(s,"comment.userName",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</b> commented 3 days ago\n                  </div>\n                  <div "),e.buffer.push(D(a.action.call(s,"deleteComment","comment",{hash:{},hashTypes:{},hashContexts:{},contexts:[s,s],types:["STRING","ID"],data:e}))),e.buffer.push(' class="comment-box__header__action">Delete</div>\n                </div>\n                <div class="comment-box__body">'),t=a._triageMustache.call(s,"comment.userComment",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n              </div>\n            </div>\n            "),n}function x(s,e){e.buffer.push('\n            <div class="center-center no-comments"><small>No comments added yet.</small></div>\n            ')}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var g,T,I,C="",N=a.helperMissing,D=this.escapeExpression,S=this;return n.buffer.push('<div class="panel variant">\n\n  <div class="panel__overlay--header variant-header">\n    <div class="list--vertical">\n      <div class="list__item">\n        <div class="list__column--small fixed center-text">\n          <div class="rank-score">'),g=a._triageMustache.call(e,"rankScore",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push("</div>\n        </div>\n        <div "),n.buffer.push(D(a.action.call(e,"toggleProperty","isShowingFullGeneName",{hash:{},hashTypes:{},hashContexts:{},contexts:[e,e],types:["STRING","STRING"],data:n}))),n.buffer.push(' class="list__column fixed">\n          '),g=a["if"].call(e,"isShowingFullGeneName",{hash:{},hashTypes:{},hashContexts:{},inverse:S.program(3,l,n),fn:S.program(1,h,n),contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push("\n          <br>\n          <small>"),g=a._triageMustache.call(e,"ensemblGeneIdString",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('</small>\n        </div>\n        <div class="list__column fixed">\n          Chr '),g=a._triageMustache.call(e,"chrom",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push(" | "),g=a._triageMustache.call(e,"startBp",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push("\n          "),g=a.unless.call(e,"isSingleBase",{hash:{},hashTypes:{},hashContexts:{},inverse:S.noop,fn:S.program(5,i,n),contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push(" |\n          "),g=a._triageMustache.call(e,"refNt",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push(" &rarr; "),g=a._triageMustache.call(e,"altNt",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push("\n\n          <br>\n\n          <small>\n            "),n.buffer.push(D((T=a.capitalize||e&&e.capitalize,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"geneAnnotation",I):N.call(e,"capitalize","geneAnnotation",I)))),n.buffer.push(" |\n            "),g=a["if"].call(e,"functionalAnnotation",{hash:{},hashTypes:{},hashContexts:{},inverse:S.program(9,p,n),fn:S.program(7,r,n),contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('\n          </small>\n        </div>\n        <div class="list__column fixed">\n          <div class="btn-group">\n            <div '),n.buffer.push(D(a.action.call(e,"openOrderModal","sangerData",{hash:{},hashTypes:{},hashContexts:{},contexts:[e,e],types:["STRING","ID"],data:n}))),n.buffer.push(' class="btn">\n              <div class="btn__body">Order Sanger</div>\n            </div>\n\n            <div class="btn">\n              <div class="btn__body">'),T=a["link-to"]||e&&e["link-to"],I={hash:{},hashTypes:{},hashContexts:{},inverse:S.noop,fn:S.program(11,c,n),contexts:[e,e],types:["STRING","ID"],data:n},g=T?T.call(e,"variants","currentFamily",I):N.call(e,"link-to","variants","currentFamily",I),(g||0===g)&&n.buffer.push(g),n.buffer.push("</div>\n            </div>\n\n            "),g=a.unless.call(e,"isHidden",{hash:{},hashTypes:{},hashContexts:{},inverse:S.program(15,o,n),fn:S.program(13,u,n),contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('\n          </div>\n          \n        </div>\n      </div>\n    </div>\n  </div>\n\n  <div class="panel__body">\n    <div class="box predictions">\n      <table class="predict">\n        <tr class="predict__top">\n          <td class="predict__main predict__part">Severity</td>\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part siftWholeExome::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"siftWholeExome",I):N.call(e,"fallback","siftWholeExome",I)))),n.buffer.push('</div>\n            <div class="predict__text">Avsift</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part polyphenDivHuman::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"polyphenDivHuman",I):N.call(e,"fallback","polyphenDivHuman",I)))),n.buffer.push('</div>\n            <div class="predict__text">PolyPhen Div</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part polyphenVarHuman::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"polyphenVarHuman",I):N.call(e,"fallback","polyphenVarHuman",I)))),n.buffer.push('</div>\n            <div class="predict__text">PolyPhen Var</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part mutationTaster::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"mutationTaster",I):N.call(e,"fallback","mutationTaster",I)))),n.buffer.push('</div>\n            <div class="predict__text">Mutation taster</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part gerpWholeExome::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"gerpWholeExome",I):N.call(e,"fallback","gerpWholeExome",I)))),n.buffer.push('</div>\n            <div class="predict__text">GERP exome</div>\n          </td>\n\n        </tr>\n\n        <tr>\n          <td class="predict__main predict__part">Frequency</td>\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part thousandG::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"thousandG",I):N.call(e,"fallback","thousandG",I)))),n.buffer.push('</div>\n            <div class="predict__text">1000 Genomes</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part esp6500::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"esp6500",I):N.call(e,"fallback","esp6500",I)))),n.buffer.push('</div>\n            <div class="predict__text">ESP 6500</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part dbsnp129::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"dbsnp129",I):N.call(e,"fallback","dbsnp129",I)))),n.buffer.push('</div>\n            <div class="predict__text">dbSNP 129</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part dbsnp::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),g=a._triageMustache.call(e,"dbsnpFlag",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('</div>\n            <div class="predict__text">dbSNP 137</div>\n          </td>\n        </tr>\n\n        <tr class="conservation-row">\n          <td class="predict__main predict__part">Conservation</td>\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part phylopWholeExome::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"phylopWholeExome",I):N.call(e,"fallback","phylopWholeExome",I)))),n.buffer.push('</div>\n            <div class="predict__text">WE PhyloP</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part gerpElement::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"gerpElement",I):N.call(e,"fallback","gerpElement",I)))),n.buffer.push('</div>\n            <div class="predict__text">GERP++ element</div>\n          </td>\n\n          <td '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{"class":":predict__part lrtWholeExome::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push('>\n            <div class="predict__number">'),n.buffer.push(D((T=a.fallback||e&&e.fallback,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"lrtWholeExome",I):N.call(e,"fallback","lrtWholeExome",I)))),n.buffer.push('</div>\n            <div class="predict__text">WE LRT</div>\n          </td>\n        </tr>\n      </table>\n    </div>\n\n    <div class="box annotations">\n      <div class="panel__wrapper">\n        <div class="panel">\n          <div class="panel__body">\n            <table class="predict annotate-table">\n              <tr>\n                <td class="predict__main predict__part">\n                  <a '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{href:"omimLink"},hashTypes:{href:"ID"},hashContexts:{href:e},contexts:[],types:[],data:n}))),n.buffer.push(' class="link" target="_blank">\n                    <div class="logo--omim">OMIM</div>\n                  </a>\n                </td>\n                <td class="predict__part">\n                  '),g=a.each.call(e,"syn","in","omim.SYNDROMS",{hash:{},hashTypes:{},hashContexts:{},inverse:S.noop,fn:S.program(17,d,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('\n                </td>\n              </tr>\n              <tr>\n                <td class="predict__main predict__part">\n                  <a href="http://www.hgmd.cf.ac.uk/ac/index.php" class="link" target="_blank">\n                    <img src="/static/img/hgmd-logo.svg">\n                  </a>\n                </td>\n                <td class="predict__part">'),g=a._triageMustache.call(e,"hgmd",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('</td>\n              </tr>\n              <tr>\n                <td class="predict__main predict__part">Disease group</td>\n                <td class="predict__part">'),g=a._triageMustache.call(e,"diseaseGroup",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('</td>\n              </tr>\n            </table>\n          </div>\n        </div>\n        <div class="panel">\n          <div class="panel__body">\n            <div class="list--vertical">\n              <div class="divider">Variant function</div>\n              '),g=a.each.call(e,"item","in","variantFunctions",{hash:{},hashTypes:{},hashContexts:{},inverse:S.program(21,b,n),fn:S.program(19,f,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('\n            </div>\n          </div>\n        </div>\n      </div>\n    </div>\n\n    <div class="box gt-call">\n      <table class="predict gt-call-table">\n        <thead>\n          <tr>\n            <th class="predict__part">IDN</th>\n            <th class="predict__part">GT</th>\n            <th class="predict__part">AD</th>\n            <th class="predict__part">PL</th>\n            <th class="predict__part">GQ</th>\n          </tr>\n        </thead>\n        <tbody>\n        '),g=a.each.call(e,"call","in","gt.gtCalls",{hash:{},hashTypes:{},hashContexts:{},inverse:S.noop,fn:S.program(23,v,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('\n\n        <tr>\n          <td class="predict__main predict__part">Inheritance models</td>\n          '),g=a.each.call(e,"model","in","geneModels",{hash:{},hashTypes:{},hashContexts:{},inverse:S.noop,fn:S.program(25,m,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push("\n        </tr>\n        </tbody>\n      </table>\n\n    </div>\n\n    "),g=a["if"].call(e,"hasCompounds",{hash:{},hashTypes:{},hashContexts:{},inverse:S.noop,fn:S.program(27,_,n),contexts:[e],types:["ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('\n\n    <div class="box list--horizontal external-links">\n      <a '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{href:"ensemblLink"},hashTypes:{href:"ID"},hashContexts:{href:e},contexts:[],types:[],data:n}))),n.buffer.push(' class="link" target="_blank">\n        <div class="link__icon"><img src="/static/img/ensembl-logo-gray.svg"></div>\n        <div class="link__text">Ensembl</div>\n      </a>\n\n      <a '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{href:"hpaLink"},hashTypes:{href:"ID"},hashContexts:{href:e},contexts:[],types:[],data:n}))),n.buffer.push(' class="link" target="_blank">\n        <div class="link__icon"><img src="/static/img/hpa-logo-gray.svg"></div>\n        <div class="link__text">HPA</div>\n      </a>\n\n      <a '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{href:"stringLink"},hashTypes:{href:"ID"},hashContexts:{href:e},contexts:[],types:[],data:n}))),n.buffer.push(' class="link" target="_blank">\n        <div class="link__icon"><img src="/static/img/string-logo-gray.svg"></div>\n        <div class="link__text">STRING</div>\n      </a>\n\n      <a '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{href:"ucscLink"},hashTypes:{href:"ID"},hashContexts:{href:e},contexts:[],types:[],data:n}))),n.buffer.push(' class="link" target="_blank">\n        <div class="link__icon"><img src="/static/img/ucsc-logo-gray.svg"></div>\n        <div class="link__text">UCSC</div>\n      </a>\n\n      <a '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{href:"entrezLink"},hashTypes:{href:"ID"},hashContexts:{href:e},contexts:[],types:[],data:n}))),n.buffer.push(' class="link" target="_blank">\n        <div class="link__icon"><img src="/static/img/entrez-logo-gray.svg"></div>\n        <div class="link__text">Entrez</div>\n      </a>\n\n      <a '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{href:"idsLink"},hashTypes:{href:"ID"},hashContexts:{href:e},contexts:[],types:[],data:n}))),n.buffer.push(' class="link" target="_blank">\n        <div class="link__icon"><img src="/static/img/scout-logo-gray.svg"></div>\n        <div class="link__text">IDs</div>\n      </a>\n\n      <a '),n.buffer.push(D(a["bind-attr"].call(e,{hash:{href:"igvLink"},hashTypes:{href:"ID"},hashContexts:{href:e},contexts:[],types:[],data:n}))),n.buffer.push(' class="link" target="_blank">\n        <div class="link__icon"><img src="/static/img/igv-logo-gray.svg"></div>\n        <div class="link__text">IGV</div>\n      </a>\n    </div>\n\n    <div class="box variant-comments">\n      <div class="panel__wrapper family-comments">\n        <div class="panel">\n          <div class="panel__overlay--header">\n            <div class="panel__overlay__title">Comments</div>\n          </div>\n          <div class="panel__body">\n            '),g=a.each.call(e,"comment","in","comments",{hash:{},hashTypes:{},hashContexts:{},inverse:S.program(31,x,n),fn:S.program(29,y,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(g||0===g)&&n.buffer.push(g),n.buffer.push('\n          </div>\n        </div>\n\n        <div class="panel">\n          <div class="panel__overlay--header">\n            <div class="panel__overlay__title">Submit new comment</div>\n          </div>\n          <div class="panel__body">\n            '),n.buffer.push(D((T=a["comment-box"]||e&&e["comment-box"],I={hash:{submit:"postComment",tags:"variantPriorities",title:"Variant comment",tagPrompt:"Set priority"},hashTypes:{submit:"STRING",tags:"ID",title:"STRING",tagPrompt:"STRING"},hashContexts:{submit:e,tags:e,title:e,tagPrompt:e},contexts:[],types:[],data:n},T?T.call(e,I):N.call(e,"comment-box",I)))),n.buffer.push("\n          </div>\n        </div>\n      </div>\n    </div>\n\n  </div>\n\n  "),n.buffer.push(D((T=a.outlet||e&&e.outlet,I={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},T?T.call(e,"modal",I):N.call(e,"outlet","modal",I)))),n.buffer.push("\n</div>"),C
-}),Ember.TEMPLATES.variants=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){var t,n,h,r="";return e.buffer.push('\n  <div class="v-panel filter-panel">\n    <div class="panel__wrapper">\n      \n      <div class="panel">\n        <div class="panel__overlay--footer filter-footer">\n          <div class="btn-group">\n            <div class="btn" '),e.buffer.push(R(a.action.call(s,"filter",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push('>\n              <div class="btn__body">Filter</div>\n            </div>\n            <div class="btn" '),e.buffer.push(R(a.action.call(s,"clinicalFilter",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push('>\n              <div class="btn__body">Clinical filter</div>\n            </div>\n            <div class="btn" '),e.buffer.push(R(a.action.call(s,"clearFilter",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["STRING"],data:e}))),e.buffer.push('>\n              <div class="btn__body">Clear</div>\n            </div>\n          </div>\n        </div>\n\n        <div class="panel__body">\n          <div class="list--vertical">\n            <div class="divider">Frequency cutoffs</div>\n            <div class="list__item">\n              <div class="list__column">\n                '),e.buffer.push(R(a.view.call(s,"Ember.RadioButton",{hash:{name:"relation",selectionBinding:"relation",value:"LESSER"},hashTypes:{name:"STRING",selectionBinding:"STRING",value:"STRING"},hashContexts:{name:s,selectionBinding:s,value:s},contexts:[s],types:["ID"],data:e}))),e.buffer.push(' Lesser\n              </div>\n\n              <div class="list__column">\n                '),e.buffer.push(R(a.view.call(s,"Ember.RadioButton",{hash:{name:"relation",selectionBinding:"relation",value:"GREATER"},hashTypes:{name:"STRING",selectionBinding:"STRING",value:"STRING"},hashContexts:{name:s,selectionBinding:s,value:s},contexts:[s],types:["ID"],data:e}))),e.buffer.push(' Greater\n              </div>\n            </div>\n\n            <div class="list__item">\n              <div class="list__column space-between">\n                <label for="genomes">1000 Genomes</label>\n                '),e.buffer.push(R((n=a.input||s&&s.input,h={hash:{id:"genomes",type:"text",valueBinding:"thousand_g",placeholder:"0.1",classNames:"text-input"},hashTypes:{id:"STRING",type:"STRING",valueBinding:"STRING",placeholder:"STRING",classNames:"STRING"},hashContexts:{id:s,type:s,valueBinding:s,placeholder:s,classNames:s},contexts:[],types:[],data:e},n?n.call(s,h):k.call(s,"input",h)))),e.buffer.push('\n              </div>\n            </div>\n\n            <div class="list__item">\n              <div class="list__column space-between">\n                <label for="dbSNP129">dbSNP 129</label>\n                '),e.buffer.push(R((n=a.input||s&&s.input,h={hash:{id:"dbSNP129",type:"text",valueBinding:"dbsnp129",placeholder:"0.1",classNames:"text-input"},hashTypes:{id:"STRING",type:"STRING",valueBinding:"STRING",placeholder:"STRING",classNames:"STRING"},hashContexts:{id:s,type:s,valueBinding:s,placeholder:s,classNames:s},contexts:[],types:[],data:e},n?n.call(s,h):k.call(s,"input",h)))),e.buffer.push('\n              </div>\n            </div>\n\n            <div class="list__item">\n              <div class="list__column space-between">\n                <label for="dbSNP132">dbSNP 132</label>\n                '),e.buffer.push(R((n=a.input||s&&s.input,h={hash:{id:"dbSNP132",type:"text",valueBinding:"dbsnp132",placeholder:"0.1",classNames:"text-input"},hashTypes:{id:"STRING",type:"STRING",valueBinding:"STRING",placeholder:"STRING",classNames:"STRING"},hashContexts:{id:s,type:s,valueBinding:s,placeholder:s,classNames:s},contexts:[],types:[],data:e},n?n.call(s,h):k.call(s,"input",h)))),e.buffer.push('\n              </div>\n            </div>\n\n            <div class="list__item">\n              <div class="list__column space-between">\n                <label for="esp6500">esp6500</label>\n                '),e.buffer.push(R((n=a.input||s&&s.input,h={hash:{id:"esp6500",type:"text",valueBinding:"esp6500",placeholder:"0.1",classNames:"text-input"},hashTypes:{id:"STRING",type:"STRING",valueBinding:"STRING",placeholder:"STRING",classNames:"STRING"},hashContexts:{id:s,type:s,valueBinding:s,placeholder:s,classNames:s},contexts:[],types:[],data:e},n?n.call(s,h):k.call(s,"input",h)))),e.buffer.push('\n              </div>\n            </div>\n\n            <div class="divider">Database and gene search</div>\n            <div class="list__item">\n              '),t=a.each.call(s,"db","in","filter.databases",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(2,l,e),contexts:[s,s,s],types:["ID","ID","ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n            </div>\n\n            <div class="list__item">\n              <div class="list__column space-between">\n                <label for="gene-name">Gene name</label>\n                '),e.buffer.push(R((n=a.input||s&&s.input,h={hash:{id:"gene-name",type:"text",valueBinding:"gene_name",placeholder:"ADK",classNames:"text-input"},hashTypes:{id:"STRING",type:"STRING",valueBinding:"STRING",placeholder:"STRING",classNames:"STRING"},hashContexts:{id:s,type:s,valueBinding:s,placeholder:s,classNames:s},contexts:[],types:[],data:e},n?n.call(s,h):k.call(s,"input",h)))),e.buffer.push("\n              </div>\n            </div>\n\n          </div>\n        </div>\n      </div> \n\n      \n      "),t=a.each.call(s,"group","in","filterGroups",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(4,i,e),contexts:[s,s,s],types:["ID","ID","ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n    </div>\n  </div>\n  "),r}function l(s,e){var t,n="";return e.buffer.push('\n              <div class="list__column">\n                '),e.buffer.push(R(a.view.call(s,"Ember.RadioButton",{hash:{name:"database",selectionBinding:"database",value:"db"},hashTypes:{name:"STRING",selectionBinding:"STRING",value:"ID"},hashContexts:{name:s,selectionBinding:s,value:s},contexts:[s],types:["ID"],data:e}))),e.buffer.push(" "),t=a._triageMustache.call(s,"db",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n              </div>\n              "),n}function i(s,e){var t,n="";return e.buffer.push('\n        <div class="panel">\n          <div class="panel__body">\n            <div class="list--vertical">\n              <div class="divider">'),t=a._triageMustache.call(s,"group.name",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n\n              "),t=a.each.call(s,"filter","in","group.filters",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(5,r,e),contexts:[s,s,s],types:["ID","ID","ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n\n            </div>\n          </div>\n        </div>\n      "),n}function r(s,e){var t,n,h,l="";return e.buffer.push('\n                <div class="list__item">\n                  <div class="list__column">\n                    '),e.buffer.push(R((n=a.input||s&&s.input,h={hash:{type:"checkbox",checkedBinding:"filter.property",classNames:"toggle-slider__input"},hashTypes:{type:"STRING",checkedBinding:"STRING",classNames:"STRING"},hashContexts:{type:s,checkedBinding:s,classNames:s},contexts:[],types:[],data:e},n?n.call(s,h):k.call(s,"input",h)))),e.buffer.push("\n                    "),t=a._triageMustache.call(s,"filter.name",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n                  </div>\n                </div>\n              "),l}function p(s,e){var t,n="";return e.buffer.push('\n\n        <span class="panel__overlay__title">\n          Variants <small>| '),t=a._triageMustache.call(s,"database",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</small>\n        </span>\n\n        "),n}function c(s,e){var t,n,h,l="";return e.buffer.push('\n\n        <div class="list__column--slim">\n          '),n=a["link-to"]||s&&s["link-to"],h={hash:{classNames:"btn"},hashTypes:{classNames:"STRING"},hashContexts:{classNames:s},inverse:E.noop,fn:E.program(10,u,e),contexts:[s],types:["STRING"],data:e},t=n?n.call(s,"index",h):k.call(s,"link-to","index",h),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n        </div>\n\n        <div class="list__column--slim">\n          Variants <small>| '),t=a._triageMustache.call(s,"database",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</small>\n        </div>\n\n        <div class="list__column--slim">\n          <div '),e.buffer.push(R(a.action.call(s,"toggleProperty","isShowingModal",{hash:{},hashTypes:{},hashContexts:{},contexts:[s,s],types:["STRING","STRING"],data:e}))),e.buffer.push(' class="btn">\n            <div class="btn__body">\n              '),t=a["if"].call(s,"isShowingModal",{hash:{},hashTypes:{},hashContexts:{},inverse:E.program(14,d,e),fn:E.program(12,o,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push(" filter\n            </div>\n          </div>\n        </div>\n\n        "),l}function u(s,e){e.buffer.push('\n            <div class="btn__body">Families</div>\n          ')}function o(s,e){e.buffer.push("Close")}function d(s,e){e.buffer.push("Open")}function f(s,e){e.buffer.push('\n        <div class="list__column--small fixed">1000 Genomes</div>\n        <div class="list__column--small fixed">PolyPhen</div>\n        <div class="list__column--small fixed">Gene annotation</div>\n        <div class="list__column fixed">Functional annotation</div>\n        <div class="list__column fixed">Inheritance models</div>\n        <div class="list__column fixed">Disease group</div>\n        <div class="list__column--small fixed">HGMD</div>\n        <div class="list__column--small fixed">Hide</div>\n        ')}function b(s,e){var t,n="";return e.buffer.push("\n      "),t=a.unless.call(s,"variant.isHidden",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(19,v,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n      "),n}function v(s,e){var t,n,h,l="";return e.buffer.push('\n      <div class="list__item">\n\n        '),n=a["link-to"]||s&&s["link-to"],h={hash:{classNames:"list__column--small fixed rank-score-column"},hashTypes:{classNames:"STRING"},hashContexts:{classNames:s},inverse:E.noop,fn:E.program(20,m,e),contexts:[s,s],types:["STRING","ID"],data:e},t=n?n.call(s,"variant","variant",h):k.call(s,"link-to","variant","variant",h),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n        <div class="list__column--small fixed">\n          <div class="hgnc-symbol">'),t=a._triageMustache.call(s,"variant.hgncSymbol",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n        </div>\n\n        "),t=a.unless.call(s,"variantLoaded",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(23,y,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n\n      </div>\n      "),l}function m(s,e){var t,n="";return e.buffer.push("\n          "),t=a._triageMustache.call(s,"variant.rankScore",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n\n          "),t=a["if"].call(s,"variant.rating",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(21,_,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n        "),n}function _(s,e){var t,n="";return e.buffer.push("\n          <div "),e.buffer.push(R(a["bind-attr"].call(s,{hash:{"class":":notification--bubble variant.rating"},hashTypes:{"class":"STRING"},hashContexts:{"class":s},contexts:[],types:[],data:e}))),e.buffer.push(">\n            "),t=a._triageMustache.call(s,"variant.rating",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n          </div>\n          "),n}function y(s,e){var t,n,h,l="";return e.buffer.push("\n        <div "),e.buffer.push(R(a["bind-attr"].call(s,{hash:{"class":":list__column--small :fixed variant.thousandG::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":s},contexts:[],types:[],data:e}))),e.buffer.push('>\n          <div class="float-left">'),e.buffer.push(R((n=a.fallback||s&&s.fallback,h={hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e},n?n.call(s,"variant.thousandG",h):k.call(s,"fallback","variant.thousandG",h)))),e.buffer.push("</div>\n\n          "),t=a["if"].call(s,"variant.isInOtherFamilies",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(24,x,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n        </div>\n\n        <div "),e.buffer.push(R(a["bind-attr"].call(s,{hash:{"class":":list__column--small :fixed variant.polyphenDivHuman::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":s},contexts:[],types:[],data:e}))),e.buffer.push(">\n          "),e.buffer.push(R((n=a.fallback||s&&s.fallback,h={hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e},n?n.call(s,"variant.polyphenDivHuman",h):k.call(s,"fallback","variant.polyphenDivHuman",h)))),e.buffer.push('\n        </div>\n\n        <div class="list__column--small fixed">'),t=a._triageMustache.call(s,"variant.geneAnnotation",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n\n        <div "),e.buffer.push(R(a["bind-attr"].call(s,{hash:{"class":":list__column :fixed variant.functionalAnnotation::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":s},contexts:[],types:[],data:e}))),e.buffer.push(">\n          "),t=a["if"].call(s,"variant.functionalAnnotation",{hash:{},hashTypes:{},hashContexts:{},inverse:E.program(28,T,e),fn:E.program(26,g,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n        </div>\n\n        <div "),e.buffer.push(R(a["bind-attr"].call(s,{hash:{"class":":list__column :fixed :scrollable variant.geneModel::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":s},contexts:[],types:[],data:e}))),e.buffer.push(">\n          "),t=a["if"].call(s,"variant.geneModels",{hash:{},hashTypes:{},hashContexts:{},inverse:E.program(28,T,e),fn:E.program(30,I,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n        </div>\n\n        <div "),e.buffer.push(R(a["bind-attr"].call(s,{hash:{"class":":list__column :fixed :scrollable variant.diseaseGroup::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":s},contexts:[],types:[],data:e}))),e.buffer.push(">\n          "),t=a["if"].call(s,"variant.diseaseGroup",{hash:{},hashTypes:{},hashContexts:{},inverse:E.program(28,T,e),fn:E.program(32,C,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n        </div>\n\n        <div "),e.buffer.push(R(a["bind-attr"].call(s,{hash:{"class":":list__column--small :fixed variant.hgmd::undef"},hashTypes:{"class":"STRING"},hashContexts:{"class":s},contexts:[],types:[],data:e}))),e.buffer.push(">\n          "),t=a["if"].call(s,"variant.hgmd",{hash:{},hashTypes:{},hashContexts:{},inverse:E.program(28,T,e),fn:E.program(34,N,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n        </div>\n\n        <div class="list__column--small fixed hide-variant">\n          <div class="btn">\n            <div '),e.buffer.push(R(a.action.call(s,"hideVariant","variant",{hash:{},hashTypes:{},hashContexts:{},contexts:[s,s],types:["STRING","ID"],data:e}))),e.buffer.push(' class="btn__body">Hide</div>\n          </div>\n        </div>\n        \n        '),l}function x(s,e){var t,n="";return e.buffer.push('\n          <div class="notification float-right">'),t=a._triageMustache.call(s,"variant.otherFamiliesCount",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n          "),n}function g(s,e){var t,n="";return e.buffer.push("\n            "),t=a._triageMustache.call(s,"variant.functionalAnnotation",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n          "),n}function T(s,e){e.buffer.push("\n            Nada\n          ")}function I(s,e){var t,n,h="";return e.buffer.push("\n            "),e.buffer.push(R((t=a["pop-over"]||s&&s["pop-over"],n={hash:{title:"variant.geneModelString",variant:"variant",show:"showPopOver",hide:"hidePopOver"},hashTypes:{title:"ID",variant:"ID",show:"STRING",hide:"STRING"},hashContexts:{title:s,variant:s,show:s,hide:s},contexts:[],types:[],data:e},t?t.call(s,n):k.call(s,"pop-over",n)))),e.buffer.push("\n          "),h}function C(s,e){var t,n="";return e.buffer.push("\n            "),t=a._triageMustache.call(s,"variant.diseaseGroup",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n          "),n}function N(s,e){var t,n="";return e.buffer.push("\n            "),t=a._triageMustache.call(s,"variant.hgmd",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n          "),n}function D(s,e){var t,n,h="";return e.buffer.push('\n      <div class="fixed-tooltip">\n        '),e.buffer.push(R((t=a.render||s&&s.render,n={hash:{},hashTypes:{},hashContexts:{},contexts:[s,s],types:["STRING","ID"],data:e},t?t.call(s,"gt-call","hoveredVariant",n):k.call(s,"render","gt-call","hoveredVariant",n)))),e.buffer.push("\n      </div>\n      "),h}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var S,G="",R=this.escapeExpression,k=a.helperMissing,E=this;return n.buffer.push("<div "),n.buffer.push(R(a["bind-attr"].call(e,{hash:{"class":"variantLoaded:panel--sidebar:panel :variants :v-panel__wrapper"},hashTypes:{"class":"STRING"},hashContexts:{"class":e},contexts:[],types:[],data:n}))),n.buffer.push(">\n\n  "),S=a["if"].call(e,"isShowingModal",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(1,h,n),contexts:[e],types:["ID"],data:n}),(S||0===S)&&n.buffer.push(S),n.buffer.push('\n\n  <div class="panel__overlay--header v-panel">\n\n    <div class="list--vertical">\n      <div class="list__item">\n        '),S=a["if"].call(e,"variantLoaded",{hash:{},hashTypes:{},hashContexts:{},inverse:E.program(9,c,n),fn:E.program(7,p,n),contexts:[e],types:["ID"],data:n}),(S||0===S)&&n.buffer.push(S),n.buffer.push('\n        \n      </div>\n\n      <div class="list__item">\n        <div class="list__column--small fixed">Rank score</div>\n        <div class="list__column--small fixed">Gene</div>\n        '),S=a.unless.call(e,"variantLoaded",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(16,f,n),contexts:[e],types:["ID"],data:n}),(S||0===S)&&n.buffer.push(S),n.buffer.push('\n      </div>\n    </div>\n  </div>\n\n  <div class="panel__body v-panel">\n    <div class="list--vertical">\n      '),S=a.each.call(e,"variant","in","model",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(18,b,n),contexts:[e,e,e],types:["ID","ID","ID"],data:n}),(S||0===S)&&n.buffer.push(S),n.buffer.push("\n\n      "),S=a["if"].call(e,"isShowingGtCall",{hash:{},hashTypes:{},hashContexts:{},inverse:E.noop,fn:E.program(36,D,n),contexts:[e],types:["ID"],data:n}),(S||0===S)&&n.buffer.push(S),n.buffer.push("\n    </div>\n  </div>\n</div>\n\n"),S=a._triageMustache.call(e,"outlet",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(S||0===S)&&n.buffer.push(S),n.buffer.push("\n"),G}),Ember.TEMPLATES["components/comment-box"]=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){var t="";return e.buffer.push("\n    "),e.buffer.push(c(a.view.call(s,"Ember.Select",{hash:{content:"tags",value:"selectedTag",optionValuePath:"content.id",optionLabelPath:"content.label",prompt:"tagPrompt"},hashTypes:{content:"ID",value:"ID",optionValuePath:"STRING",optionLabelPath:"STRING",prompt:"ID"},hashContexts:{content:s,value:s,optionValuePath:s,optionLabelPath:s,prompt:s},contexts:[s],types:["ID"],data:e}))),e.buffer.push("\n    "),t}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var l,i,r,p="",c=this.escapeExpression,u=a.helperMissing,o=this;return n.buffer.push('<div class="comment-box">\n  '),n.buffer.push(c((i=a.textarea||e&&e.textarea,r={hash:{valueBinding:"body",classNames:"comment-box__body--input",placeholder:"Write your words here..."},hashTypes:{valueBinding:"STRING",classNames:"STRING",placeholder:"STRING"},hashContexts:{valueBinding:e,classNames:e,placeholder:e},contexts:[],types:[],data:n},i?i.call(e,r):u.call(e,"textarea",r)))),n.buffer.push('\n  <div class="comment-box__footer">\n    <div class="comment-box__footer__body">'),l=a._triageMustache.call(e,"title",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(l||0===l)&&n.buffer.push(l),n.buffer.push("</div>\n    "),l=a["if"].call(e,"hasTags",{hash:{},hashTypes:{},hashContexts:{},inverse:o.noop,fn:o.program(1,h,n),contexts:[e],types:["ID"],data:n}),(l||0===l)&&n.buffer.push(l),n.buffer.push("\n    <div "),n.buffer.push(c(a.action.call(e,"submit",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["STRING"],data:n}))),n.buffer.push(' class="comment-box__submit">Submit</div>\n  </div>\n</div>\n'),p}),Ember.TEMPLATES["components/modal-dialog"]=Ember.Handlebars.template(function(s,e,a,t,n){this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var h,l="",i=this.escapeExpression;return n.buffer.push('<div class="modal__overlay" '),n.buffer.push(i(a.action.call(e,"close",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["STRING"],data:n}))),n.buffer.push('>\n  <div class="modal__wrapper">\n    <div class="modal" '),n.buffer.push(i(a.action.call(e,{hash:{bubbles:!1},hashTypes:{bubbles:"BOOLEAN"},hashContexts:{bubbles:e},contexts:[],types:[],data:n}))),n.buffer.push(">\n      "),h=a._triageMustache.call(e,"yield",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(h||0===h)&&n.buffer.push(h),n.buffer.push("\n    </div>\n  </div>\n</div>"),l}),Ember.TEMPLATES["components/pop-over"]=Ember.Handlebars.template(function(s,e,a,t,n){this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var h,l="";return h=a._triageMustache.call(e,"title",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(h||0===h)&&n.buffer.push(h),n.buffer.push("\n"),l}),Ember.TEMPLATES["family/index"]=Ember.Handlebars.template(function(s,e,a,t,n){function h(s,e){e.buffer.push("Variants")}function l(s,e){e.buffer.push("\n                Show raw pedigree\n                ")}function i(s,e){e.buffer.push("\n                Hide raw pedigree\n                ")}function r(s,e){var t="";return e.buffer.push('\n    <div class="box raw-pedigree">\n      '),e.buffer.push(y(a._triageMustache.call(s,"family.pedigree",{hash:{unescaped:"true"},hashTypes:{unescaped:"STRING"},hashContexts:{unescaped:s},contexts:[s],types:["ID"],data:e}))),e.buffer.push("\n    </div>\n\n    "),t}function p(s,e){var t,n,h,l="";return e.buffer.push('\n    <div class="box family">\n      <table class="predict gt-call-table">\n\n        '),t=a.each.call(s,"sample","in","family.samples",{hash:{},hashTypes:{},hashContexts:{},inverse:x.noop,fn:x.program(10,c,e),contexts:[s,s,s],types:["ID","ID","ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n      </table>\n    </div>\n\n    "),t=a["if"].call(s,"hasGeneModels",{hash:{},hashTypes:{},hashContexts:{},inverse:x.noop,fn:x.program(12,u,e),contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n\n    <div class="box panel__wrapper family-comments">\n      <div class="panel">\n        <div class="panel__overlay--header">\n          <div class="panel__overlay__title">Diagnostic comments</div>\n        </div>\n        <div class="panel__body">\n          '),t=a.each.call(s,"comment","in","diagnosticComments",{hash:{},hashTypes:{},hashContexts:{},inverse:x.program(17,f,e),fn:x.program(15,d,e),contexts:[s,s,s],types:["ID","ID","ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n        </div>\n      </div>\n\n      <div class="panel">\n        <div class="panel__overlay--header">\n          <div class="panel__overlay__title">Research comments</div>\n        </div>\n        <div class="panel__body">\n\n          '),t=a.each.call(s,"comment","in","researchComments",{hash:{},hashTypes:{},hashContexts:{},inverse:x.program(17,f,e),fn:x.program(15,d,e),contexts:[s,s,s],types:["ID","ID","ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('\n\n        </div>\n      </div>\n    </div>\n\n    <div class="box panel__wrapper family-submit-comment">\n\n      <div class="panel">\n        <div class="panel__body">\n          '),e.buffer.push(y((n=a["comment-box"]||s&&s["comment-box"],h={hash:{submit:"postComment",tags:"commentCategories",type:"IEM",title:"IEM Comment"},hashTypes:{submit:"STRING",tags:"ID",type:"STRING",title:"STRING"},hashContexts:{submit:s,tags:s,type:s,title:s},contexts:[],types:[],data:e},n?n.call(s,h):g.call(s,"comment-box",h)))),e.buffer.push('\n        </div>\n      </div>\n\n      <div class="panel">\n        <div class="panel__body">\n          '),e.buffer.push(y((n=a["comment-box"]||s&&s["comment-box"],h={hash:{submit:"postComment",tags:"commentCategories",type:"research",title:"Research Comment"},hashTypes:{submit:"STRING",tags:"ID",type:"STRING",title:"STRING"},hashContexts:{submit:s,tags:s,type:s,title:s},contexts:[],types:[],data:e},n?n.call(s,h):g.call(s,"comment-box",h)))),e.buffer.push("\n        </div>\n      </div>\n\n    </div>\n    "),l}function c(s,e){var t,n="";return e.buffer.push('\n          <tr>\n            <td class="predict__main predict__part">'),t=a._triageMustache.call(s,"sample.id",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</td>\n            <td class="predict__part affected">\n              \n              <div class="family-member">'),t=a._triageMustache.call(s,"sample.gender",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push(" "),t=a._triageMustache.call(s,"sample.type",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push('</div>\n            </td>\n            <td class="predict__part">'),t=a._triageMustache.call(s,"sample.captureKit",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</td>\n          </tr>\n        "),n}function u(s,e){var t,n="";return e.buffer.push('\n    <div class="box">\n      <table class="predict">\n        <tr>\n          <td class="predict__main predict__part">Possible gene models</td>\n          '),t=a.each.call(s,"model","in","family.samples.1.inheritanceModels",{hash:{},hashTypes:{},hashContexts:{},inverse:x.noop,fn:x.program(13,o,e),contexts:[s,s,s],types:["ID","ID","ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("\n        </tr>\n      </table>\n    </div>\n    "),n}function o(s,e){var t,n="";return e.buffer.push('\n            <td class="predict__part">'),t=a._triageMustache.call(s,"model",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</td>\n          "),n}function d(s,e){var t,n="";return e.buffer.push('\n          <div class="comment-box__wrapper">\n            <div class="comment-box">\n              <div class="comment-box__header">\n                <div class="comment-box__header__body">\n                  <b>'),t=a._triageMustache.call(s,"comment.userName",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</b> commented 3 days ago\n                </div>\n                <div "),e.buffer.push(y(a.action.call(s,"deleteComment","comment",{hash:{},hashTypes:{},hashContexts:{},contexts:[s,s],types:["STRING","ID"],data:e}))),e.buffer.push(' class="comment-box__header__action">Delete</div>\n              </div>\n              <div class="comment-box__body">'),t=a._triageMustache.call(s,"comment.userComment",{hash:{},hashTypes:{},hashContexts:{},contexts:[s],types:["ID"],data:e}),(t||0===t)&&e.buffer.push(t),e.buffer.push("</div>\n            </div>\n          </div>\n          "),n}function f(s,e){e.buffer.push('\n          <div class="center-center no-comments"><small>No comments added yet.</small></div>\n          ')}this.compilerInfo=[4,">= 1.0.0"],a=this.merge(a,Ember.Handlebars.helpers),n=n||{};var b,v,m,_="",y=this.escapeExpression,x=this,g=a.helperMissing;return n.buffer.push('<div class="panel family">\n\n  <div class="panel__overlay--header family-header">\n\n    <div class="list--vertical">\n      <div class="list__item">\n        <div class="list__column fixed center-text">\n          <div class="family-id">Family '),b=a._triageMustache.call(e,"family.id",{hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n}),(b||0===b)&&n.buffer.push(b),n.buffer.push('</div>\n        </div>\n        <div class="list__column fixed center-text">\n          Last update: '),n.buffer.push(y((v=a.fromNow||e&&e.fromNow,m={hash:{},hashTypes:{},hashContexts:{},contexts:[e],types:["ID"],data:n},v?v.call(e,"family.analyzedDate",m):g.call(e,"fromNow","family.analyzedDate",m)))),n.buffer.push('\n        </div>\n        <div class="list__column fixed">\n          <div class="btn-group">\n            <div class="btn">\n              <div class="btn__body">\n                '),v=a["link-to"]||e&&e["link-to"],m={hash:{},hashTypes:{},hashContexts:{},inverse:x.noop,fn:x.program(1,h,n),contexts:[e,e],types:["STRING","ID"],data:n},b=v?v.call(e,"variants","family",m):g.call(e,"link-to","variants","family",m),(b||0===b)&&n.buffer.push(b),n.buffer.push("\n              </div>\n            </div>\n\n            <div "),n.buffer.push(y(a.action.call(e,"toggleProperty","isShowingRawPedigree",{hash:{},hashTypes:{},hashContexts:{},contexts:[e,e],types:["STRING","STRING"],data:n}))),n.buffer.push(' class="btn">\n              <div class="btn__body">\n                '),b=a.unless.call(e,"isShowingRawPedigree",{hash:{},hashTypes:{},hashContexts:{},inverse:x.program(5,i,n),fn:x.program(3,l,n),contexts:[e],types:["ID"],data:n}),(b||0===b)&&n.buffer.push(b),n.buffer.push('\n              </div>\n            </div>\n          </div>\n        </div>\n      </div>\n    </div>\n  </div>\n\n  <div class="panel__body family-body">\n\n    '),b=a["if"].call(e,"isShowingRawPedigree",{hash:{},hashTypes:{},hashContexts:{},inverse:x.program(9,p,n),fn:x.program(7,r,n),contexts:[e],types:["ID"],data:n}),(b||0===b)&&n.buffer.push(b),n.buffer.push("\n\n  </div>\n</div>\n"),_});
